% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{anova.droprate}
\alias{anova.droprate}
\title{Compares model fits of two \code{droprate} objects}
\usage{
\method{anova}{droprate}(object, y, conf.level = 0.95, digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{droprate}, usually a result of a call to \link[droprate]{estinterval}}

\item{y}{an object of class \code{droprate}, usually a result of a call to \link[droprate]{estinterval}}

\item{conf.level}{confidence level for the deviance test}

\item{digits}{the number of digits for printing to screen}

\item{...}{other arguments to be passed to low level functions}
}
\value{
A list of class "\code{anova.droprate}" with the best model (1 or 2), deviance statistic and test results
}
\description{
Compare model fits of two \code{droprate} objects estimated on the same data
}
\examples{
data(goosedrop)
model1=estinterval(goosedrop$interval,fun="normal")
model2=estinterval(goosedrop$interval,fun="gamma")
# the model assuming normally distributed intervals outperforms
# the model assuming gamma-distributed intervals:
anova(model1,model2)
# visually inspect model2 fit:
plot(model2)
# The observed distribution has intervals near zero.
# We add a small random baseline to reduce the effect
# of intervals near zero on the fit result.
model3=estinterval(goosedrop$interval,fun="gamma",fpp=.05)
# model3 performs as good as model1:
anova(model1,model3)
}

