% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\docType{package}
\name{droprate-package}
\alias{droprate}
\alias{droprate-package}
\title{Analyse interval data with missed event observations}
\description{
Droprate allows you to calculate mean event intervals
and compare means and variances of groups of interval data,
while taking into account missed event observations
}
\details{
The central function of package \pkg{droprate} is
\code{\link{estinterval}}, which can be used to estimate the
mean event interval (and its standard deviation) from interval
data with missed events. This is
achieved by fitting the theoretical probability density
\code{\link{intervalpdf}} to the interval data

This package was designed initially for
analysing dropping intervals of grazing geese, but can be used to analyse any
kind of interval data derived from distinct event observations.
One interval is defined as the time between observing a dropping be excreted, until the time the
next dropping is excreted. The package provides a way of taking into account missed observations
(excreted droppings), which leads to occasional observed intervals at integer multiples of the
true dropping interval.

Sets of interval data can be compared using \code{\link{ttest}} and \code{\link{vartest}}

The package comes with a test dataset \code{\link{goosedrop}}
}
\references{
Bédard, J. & Gauthier, G. (1986) Assessment of faecal output in geese. Journal of Applied Ecology, 23, 77–90.

Dokter et al. 2016 ???
}

