% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{droprate}
\alias{droprate}
\title{Analyse event rate with missed event observations}
\usage{
droprate(data, mu = mean(data), sigma = sd(data), p = 0.2, N = 5,
  fun = "normal")
}
\arguments{
\item{data}{A numeric list of intervals.}

\item{mu}{Start value for the numeric optimization for the mean interval rate.}

\item{sigma}{Start value for the numeric optimization for the standard deviation of the interval rate.}

\item{p}{Start value for the numeric optimization for the chance to miss an observation of an event.}

\item{N}{Maximum number of missed observations to be taken into account.}

\item{fun}{Assumed distribution for the intervals, one of "\code{normal}" or "\code{gamma}", corresponding
to the \link[stats]{Normal} and \link[stats]{GammaDist} distributions}
}
\value{
This function returns an object of class \code{droprate}
}
\description{
Analyse event rate with missed event observations
}
\examples{
data(terschelling)
dr=droprate(terschelling$interval)
plot(dr)
summary(dr)
}

