% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{fundamental}
\alias{fundamental}
\title{Estimate which intervals are fundamental}
\usage{
fundamental(x, conf.level = 0.9)
}
\arguments{
\item{x}{object inheriting from class \code{intRval}, usually a result of a call to \link[intRval]{estinterval}}

\item{conf.level}{confidence level for identifying intervals as fundamental}
}
\value{
logical atomic vector of the same length as \code{x$data}
}
\description{
Estimates which intervals in a dataset are fundamental intervals, i.e. an
interval not containing a missed arrival observation
}

