% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\docType{package}
\name{intRval-package}
\alias{intRval}
\alias{intRval-package}
\title{Analyse interval data with missed arrival observations}
\description{
\pkg{intRval} calculates means and variances of arrival intervals
corrected for missed arrival observations, and compares means
 and variances of groups of interval data.
}
\details{
The central function of package \pkg{intRval} is
\code{\link{estinterval}}, which can be used to estimate the
mean arrival interval (and its standard deviation) from interval
data with missed arrivals. This is
achieved by fitting the theoretical probability density
\code{\link{intervalpdf}} to the interval data

The package was designed originally for analysing dropping intervals
of grazing geese to estimate their faecal output, but can be used to analyse general
interval data where intervals are derived from distinct arrival observations.

Intervals are defined as the time between observed arrival events (e.g. the time between one excreted droppings to the next)
The package provides a way of taking into account missed observations
(excreted droppings), which lead to occasional observed intervals at integer multiples of the
true arrival interval.

Sets of interval data can be compared using \code{\link{ttest}} and \code{\link{vartest}}

The package comes with a test dataset \code{\link{goosedrop}}
}
\references{
Dokter et al. 2016 ???

B\'{e}dard, J. & Gauthier, G. (1986) Assessment of faecal output in geese. Journal of Applied Ecology, 23, 77-90.

Owen, M. 1971. The Selection of Feeding Site by White-Fronted Geese in Winter. Journal of Applied Ecology 8: 905-917.
}

