% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{partition}
\alias{partition}
\title{Estimate within-group variation}
\usage{
partition(x, conf.level = 0.9, alpha = 0.05, silent = F)
}
\arguments{
\item{x}{object inheriting from class \code{intRval}}

\item{conf.level}{confidence level passed to function \link[intRval]{fundamental}, used in selecting fundamental intervals}

\item{alpha}{significance level for differences within and between groups or subjects}

\item{silent}{logical, if \code{TRUE} print no text to console}
}
\value{
A list with results
\describe{
  \item{\code{sigma.within}}{within-group standard deviation in interval length, estimated on fundamental intervals with repeated measures only}
  \item{\code{sigma}}{the total standard deviation in interval length, copied from \code{x$sigma}}
  \item{\code{p.within}}{p-value form a likelihood-ratio test indicating whether there is evidence for a random effect of group or subject}
  \item{\code{n.within}}{average number of intervals per group}
  \item{\code{n.total}}{total number of intervals}
  \item{\code{n.repeat}}{number of fundamental intervals with repeated measures, the size of the dataset on which \code{sigma.within} was estimated}
  \item{\code{p<alpha}}{logical. Whether there was significant evidence for a difference in within- and between-group/subject variance}
}
}
\description{
Estimate within-group variation in interval length
}
\examples{
# select the group of intervals observed on Terschelling island
dropset=goosedrop[goosedrop$site=="terschelling",]
# estimate an interval model, with separate within- and between-group variation:
dr=estinterval(data=dropset$interval,group = dropset$bout_id)
# plot the model fit:
plot(dr)
# estimate within-group variation, and its significance:
output=partition(dr)
# print within-group standard deviation:
output$sigma.within
# is the model including within-group standard deviation signicant,
# relative to a null model without separate within-group sd,
# at the specified confidence level \\code{alpha}?
output$`p<alpha`  #> TRUE
}
