% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.anova.intRval}
\alias{print.anova.intRval}
\title{print method for class \code{anova.intRval}}
\usage{
\method{print}{anova.intRval}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{An object of class \code{anova.intRval}, usually a result of a call to \link[intRval]{anova.intRval}}
}
\description{
print method for class \code{anova.intRval}
}
\keyword{internal}
