% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.anova.intRvals}
\alias{print.anova.intRvals}
\title{print method for class \code{anova.intRvals}}
\usage{
\method{print}{anova.intRvals}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{An object of class \code{anova.intRvals}, usually a result of a call to \link[intRvals]{anova.intRvals}}
}
\description{
print method for class \code{anova.intRvals}
}
\keyword{internal}
