% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.droprate}
\alias{print.droprate}
\title{print method for class \code{droprate}}
\usage{
\method{print}{droprate}(x, digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{x}{An object of class \code{droprate}, usually a result of a call to \link[droprate]{droprate}}
}
\description{
print method for class \code{droprate}
}
\keyword{internal}

