% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.intRval}
\alias{print.intRval}
\title{print method for class \code{intRval}}
\usage{
\method{print}{intRval}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{intRval}, usually a result of a call to \link[intRval]{estinterval}}
}
\description{
print method for class \code{intRval}
}
\keyword{internal}

