% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.intRvals}
\alias{print.intRvals}
\title{print method for class \code{intRvals}}
\usage{
\method{print}{intRvals}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{intRvals}, usually a result of a call to \link[intRvals]{estinterval}}
}
\description{
print method for class \code{intRvals}
}
\keyword{internal}
