% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.summary.droprate}
\alias{print.summary.droprate}
\title{print method for class \code{summary.droprate}}
\usage{
\method{print}{summary.droprate}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{An object of class \code{symmary.droprate}, usually a result of a call to \link[droprate]{summary.droprate}}
}
\description{
print method for class \code{summary.droprate}
}
\keyword{internal}

