% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.summary.intRval}
\alias{print.summary.intRval}
\title{print method for class \code{summary.intRval}}
\usage{
\method{print}{summary.intRval}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{An object of class \code{symmary.intRval}, usually a result of a call to \link[intRval]{summary.intRval}}
}
\description{
print method for class \code{summary.intRval}
}
\keyword{internal}

