% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{print.summary.intRvals}
\alias{print.summary.intRvals}
\title{print method for class \code{summary.intRvals}}
\usage{
\method{print}{summary.intRvals}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{An object of class \code{symmary.intRvals}, usually a result of a call to \link[intRvals]{summary.intRvals}}
}
\description{
print method for class \code{summary.intRvals}
}
\keyword{internal}
