% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\docType{data}
\name{terschelling}
\alias{terschelling}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 571 rows and 2 columns.}
\usage{
terschelling
}
\description{
This is data to be included in my package
}
\author{
Adriaan Dokter \email{a.m.dokter@uva.nl}
}
\references{
\url{data_blah.com}
}
\keyword{datasets}

