% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{droprate}
\alias{droprate}
\title{Analyse event rate with missed event observations}
\usage{
droprate(data, mu = mean(data), sigma = sd(data)/2, p = 0.2, N = 5,
  fun = "normal")
}
\arguments{
\item{data}{A numeric list of intervals.}

\item{mu}{Start value for the numeric optimization for the mean interval rate.}

\item{sigma}{Start value for the numeric optimization for the standard deviation of the interval rate.}

\item{p}{Start value for the numeric optimization for the chance to miss an observation of an event.}

\item{N}{Maximum number of missed observations to be taken into account.}

\item{fun}{Assumed distribution for the intervals, one of "\code{normal}" or "\code{gamma}", corresponding
to the \link[stats]{Normal} and \link[stats]{GammaDist} distributions}
}
\value{
This function returns an object of class \code{droprate}
}
\description{
Analyse event rate with missed event observations
}
\examples{
data(goosedrop)
dr=droprate(goosedrop$interval)
plot(dr)
summary(dr)

# let's analyze dropping intervals by site
dr1=droprate(goosedrop[goosedrop$site=="schiermonnikoog",]$interval)
dr2=droprate(goosedrop[goosedrop$site=="terschelling",]$interval)

# The T test that accounts for missed observations shows the
# mean dropping interval is not significantly different at the two sites:
ttest(dr1,dr2)
# not accounting for missed observations leads to the false
# conclusion that the mean intervals are significantly different:
t.test(dr1$data,dr2$data)

}

