% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{estinterval}
\alias{estinterval}
\title{Estimate interval mean and variance accounting for missed arrival observations}
\usage{
estinterval(data, mu = median(data), sigma = sd(data)/2, p = 0.2,
  N = 5L, fun = "gamma", trunc = c(0, Inf), fpp = (if (fpp.method ==
  "fixed") 0 else 0.1), fpp.method = "fixed", p.method = "auto",
  conf.level = 0.95, ...)
}
\arguments{
\item{data}{A numeric list of intervals.}

\item{mu}{Start value for the numeric optimization for the mean arrival rate.}

\item{sigma}{Start value for the numeric optimization for the standard deviation of the arrival rate.}

\item{p}{Start value for the numeric optimization for the probability to miss an observation of an arrival.}

\item{N}{Maximum number of missed observations to be taken into account (default N=5).}

\item{fun}{Assumed distribution for the intervals, one of "\code{normal}" or "\code{gamma}", corresponding
to the \link[stats]{Normal} and \link[stats]{GammaDist} distributions}

\item{trunc}{Use a truncated probability density function with range \code{trunc}}

\item{fpp}{Baseline proportion of intervals distributed as a random poisson process with mean arrival rate \code{mu}}

\item{fpp.method}{A string equal to 'fixed' or 'auto'. When 'auto' \code{fpp} is optimized as a free model parameter,
in which case \code{fpp} is taken as start value in the optimisation}

\item{p.method}{A string equal to 'fixed' or 'auto'. When 'auto' \code{p} is optimized as a free model parameter,
in which case \code{p} is taken as start value in the optimisation}

\item{conf.level}{Confidence level for deviance test that checks whether model with nonzero miss chance
\code{p} significantly outperforms a model without a miss chance (\code{p=0}).}

\item{...}{Additional arguments to be passed to \link[stats]{optim}}
}
\value{
This function returns an object of class \code{droprate}, which is a list containing the following:
\describe{
  \item{\code{data}}{the interval data}
  \item{\code{mean}}{the modelled mean interval}
  \item{\code{stdev}}{the modelled interval standard deviation}
  \item{\code{p}}{the modelled probability to not observe an arrival}
  \item{\code{fpp}}{the modelled fraction of arrivals following a random poisson process, see \link[droprate]{intervalpdf}}
  \item{\code{N}}{the highest number of consecutive missed arrivals taken into account, see \link[droprate]{intervalpdf}}
  \item{\code{convergence}}{convergence field of \link[stats]{optim}}
  \item{\code{counts}}{counts field of \link[stats]{optim}}
  \item{\code{loglik}}{vector of length 2, with first element the log-likelihood of the fitted model, and second element the log-likelihood of the model without a miss chance (i.e. \code{p}=0)}
  \item{\code{df.residual}}{degrees of freedom, a 2-vector (1, number of intervals - \code{n.param})}
  \item{\code{n.param}}{number of optimized model parameters}
  \item{\code{distribution}}{assumed interval distribution, one of 'gamma' or 'normal'}
  \item{\code{trunc}}{interval range over which the interval pdf was truncated and normalized}
  \item{\code{fpp.method}}{A string equal to 'fixed' or 'auto'. When 'auto' \code{fpp} has been optimized as a free model parameter}
  \item{\code{p.method}}{A string equal to 'fixed' or 'auto'. When 'auto' \code{f} has been optimized as a free model parameter}
}
}
\description{
Estimate interval mean and variance accounting for missed arrival observations
}
\details{
The probability density function for observed intervals \link[droprate]{intervalpdf}
is fit to \code{data} by maximization of the
associated log-likelihood using \link[stats]{optim}.
}
\examples{
data(goosedrop)
dr=estinterval(goosedrop$interval)
plot(dr)
summary(dr)

# let's estimate mean and variance of dropping intervals by site
dr1=estinterval(goosedrop[goosedrop$site=="schiermonnikoog",]$interval)
dr2=estinterval(goosedrop[goosedrop$site=="terschelling",]$interval)
# The T test that accounts for missed observations shows that the
# mean dropping interval are significantly different at the two sites:
ttest(dr1,dr2)
# assuming gamma-distributed intervals, analysing period 5 only
# assuming a fraction of intervals to be distributed randomly (fpp='auto'),
dr3=estinterval(goosedrop[goosedrop$site=="schiermonnikoog" & goosedrop$period==5,]$interval,
fpp.method="auto",fun="gamma")
dr4=estinterval(goosedrop[goosedrop$site=="terschelling" & goosedrop$period==5,]$interval,
fpp.method="auto",fun="gamma")
# plot the fits:
plot(dr3,xlim=c(0,1000))
plot(dr4,xlim=c(0,1000))
# mean dropping interval are not significantly different
# at the two sites (on a 0.95 confidence level):
ttest(dr3,dr4)
# not accounting for missed observations leads to a (spurious)
# larger difference in means, which also increases
# the apparent statistical significance of the difference of means
t.test(dr3$data,dr4$data)

}

