% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{estinterval}
\alias{estinterval}
\title{Estimate interval mean and variance accounting for missed arrival observations}
\usage{
estinterval(data, mu = mean(data), sigma = sd(data)/2, p = 0.2, N = 5,
  fun = "normal")
}
\arguments{
\item{data}{A numeric list of intervals.}

\item{mu}{Start value for the numeric optimization for the mean interval rate.}

\item{sigma}{Start value for the numeric optimization for the standard deviation of the interval rate.}

\item{p}{Start value for the numeric optimization for the chance to miss an observation of an arrival.}

\item{N}{Maximum number of missed observations to be taken into account (default N=5).}

\item{fun}{Assumed distribution for the intervals, one of "\code{normal}" or "\code{gamma}", corresponding
to the \link[stats]{Normal} and \link[stats]{GammaDist} distributions}
}
\value{
This function returns an object of class \code{droprate}
}
\description{
Estimate interval mean and variance accounting for missed arrival observations
}
\details{
The probability density function for observed intervals \link[droprate]{intervalpdf}
is fit to \code{data} by maximization of the
associated log-likelihood using \link[stats]{optim}.
}
\examples{
data(goosedrop)
dr=estinterval(goosedrop$interval)
plot(dr)
summary(dr)

# let's estimate mean and variance of dropping intervals by site
dr1=estinterval(goosedrop[goosedrop$site=="schiermonnikoog",]$interval)
dr2=estinterval(goosedrop[goosedrop$site=="terschelling",]$interval)

# The T test that accounts for missed observations shows that the
# mean dropping interval is not significantly different at the two sites:
ttest(dr1,dr2)
# not accounting for missed observations leads to the false
# conclusion that the mean intervals are significantly different:
t.test(dr1$data,dr2$data)

}

