% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{intervalpdf}
\alias{intervalpdf}
\title{Probability density function (PDF) of an observed interval distribution}
\usage{
intervalpdf(data = seq(0, 1000), mu = 200, sigma = 40, p = 0.3, N = 5,
  fun = "normal")
}
\arguments{
\item{data}{A list of intervals for which to calculate the probability density}

\item{mu}{The mean of the true interval distribution}

\item{sigma}{The standard deviation of the true interval distribution}

\item{p}{The chance that the event that defines the start or end of an interval is not observed}

\item{N}{The maximum number of consecutive missed events to take into consideration}

\item{fun}{assumed distribution family of the true interval distribution, one of
 "\code{normal}" or "\code{gamma}", corresponding
to the \link[stats]{Normal} and \link[stats]{GammaDist} distributions.}
}
\value{
This function returns a list with data, corresponding to the model fit
}
\description{
Observed intervals are assumed to be sampled through observation of continuous
distinct events in time. Two subsequently observed events mark the start and end
of an interval. The chance that an event is not observed can be nonzero, leading to
observed intervals at integer multiples of the true interval.
}
\details{
By default intervals x are assumed to follow a normal distribution \eqn{N(\mu,\sigma^2)}{N(\mu,\sigma^2)},
with a probability density function \eqn{\phi(x)}:
\deqn{\phi(x|\mu,\sigma^2)~N(\mu,\sigma^2)}{\phi(x|\mu,\sigma^2)~N(\mu,\sigma^2)}
with \eqn{\mu} the average event interval and \eqn{\sigma} its associated standard deviation.
The probability density function \eqn{\phi}obs of observed event intervals
in a scenario where the chance to not observe an event is nonzero,
will be a superposition of several normal distributions, at multiples of the fundamental mean
event interval. Normal distribution \eqn{i} will correspond to those intervals where \eqn{i} events have been
missed consecutively. If \eqn{p} equals this chance of not observing a dropping, then the
probability \eqn{P(i)} to miss \eqn{i} consecutive droppings equals
\deqn{P(i)=p^i-p^{i+1}}{P(i)=p^i-p^{i+1}}
The width of normal distribution i will be broadened relative to the fundamental, according to
standard uncertainty propagation in the case of addition, such that we may write for the observed PDF, \eqn{\phi{obs}}:
\deqn{\phi_{obs}(x | \mu, \sigma^2,p)=\sum_{i=1}^\infty P(i-1) \phi(x | i \mu,i \sigma^2)}{\phi{obs}(x | \mu,\sigma^2,p)=\sum_{i=1}^\infty P(i-1) \phi(x | i \mu,i \sigma^2)}
In practice, this probability density function is well approximate when the infinite sum is capped at a finite integer N.
Be default the sum is ran up to N=5.
}
\examples{
# a low miss chance results in an observed PDF
# with primarily a single peak, with a mean and standard
# deviation almost identical to the true interval
# distribution:
plot(intervalpdf(mu=200,sigma=40,p=0.01),type='l',col='red')

# a higher miss chance results in an observed PDF with multiple
#  peaks at integer multiples of the mean of the true
# interval distribution
plot(intervalpdf(mu=200,sigma=40,p=0.4),type='l',col='red')
}

