% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{summary.droprate}
\alias{summary.droprate}
\title{summary method for class \code{droprate}}
\usage{
\method{summary}{droprate}(x)
}
\arguments{
\item{x}{An object of class \code{droprate}, usually a result of a call to \link[droprate]{droprate}}
}
\value{
The function \code{summary.droprate} computes and returns a list of summary statistics
}
\description{
summary method for class \code{droprate}
}
\examples{
data(terschelling)
dr=droprate(terschelling$interval)
summary(dr)
}

