% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DroppingInterval.R
\name{summary.droprate}
\alias{summary.droprate}
\title{summary method for class \code{droprate}}
\usage{
\method{summary}{droprate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{droprate}, usually a result of a call to \link[droprate]{estinterval}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.droprate} computes and returns a list of summary statistics
}
\description{
summary method for class \code{droprate}
}
\examples{
data(goosedrop)
dr=estinterval(goosedrop$interval)
summary(dr)
}

